chore: convert deny-list integ test to integ-runner
