+++ .eslintrc.json
@@ -51,8 +51,8 @@
       "error",
       {
         "devDependencies": [
+          "**/test/**",
           "**/src/__tests__/**",
-          "**/build-tools/**",
           "src/**/*.lambda.ts",
           "src/**/*.lambda-shared.ts",
           "src/**/*.ecstask.ts",
