+++ .gitattributes
@@ -28,9 +28,9 @@
 /lib/backend/transliterator/transliterator.ecs-entrypoint.bundle/Dockerfile linguist-generated
 /LICENSE linguist-generated
 /package.json linguist-generated
-/src/__tests__/backend/deny-list/integ/catalog-builder-mock.ts linguist-generated
-/src/__tests__/backend/deny-list/integ/trigger.client-test.ts linguist-generated
-/src/__tests__/backend/deny-list/integ/trigger.prune-test.ts linguist-generated
+/src/__tests__/backend/deny-list/mocks/catalog-builder-mock.ts linguist-generated
+/src/__tests__/backend/deny-list/mocks/trigger.client-test.ts linguist-generated
+/src/__tests__/backend/deny-list/mocks/trigger.prune-test.ts linguist-generated
 /src/backend/catalog-builder/catalog-builder.ts linguist-generated
 /src/backend/deny-list/prune-handler.ts linguist-generated
 /src/backend/deny-list/prune-queue-handler.ts linguist-generated
