+++ .gitignore
@@ -57,8 +57,8 @@ tsconfig.json
 !/.github/workflows/upgrade-cdklabs-projen-project-types-main.yml
 !/.github/workflows/upgrade-main.yml
 !/.github/workflows/upgrade-dev-deps-main.yml
-!/test/fixtures/tests/package.tgz
-/test/integ.transliterator.ecstask.ts.snapshot/asset.*
+!/test/fixtures/**
+/test/*.snapshot/asset.*
 !/src/third-party-types/*
 /website
 .vscode/
@@ -90,9 +90,9 @@ lib/__tests__/devapp/cdk.out
 !/src/backend/deny-list/prune-queue-handler.ts
 !/src/backend/deny-list/prune-handler.ts
 !/src/backend/catalog-builder/catalog-builder.ts
-!/src/__tests__/backend/deny-list/integ/trigger.prune-test.ts
-!/src/__tests__/backend/deny-list/integ/trigger.client-test.ts
-!/src/__tests__/backend/deny-list/integ/catalog-builder-mock.ts
+!/src/__tests__/backend/deny-list/mocks/trigger.prune-test.ts
+!/src/__tests__/backend/deny-list/mocks/trigger.client-test.ts
+!/src/__tests__/backend/deny-list/mocks/catalog-builder-mock.ts
 !/src/backend/transliterator/transliterator.ts
 !/src/backend/transliterator/transliterator.ecs-entrypoint.ts
 !/lib/backend/transliterator/transliterator.ecs-entrypoint.bundle/Dockerfile
@@ -109,18 +109,5 @@ src/__tests__/domain-redirect.integ.cdkout/**/.cache
 src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy
 src/__tests__/.tmp.domain-redirect.integ.cdkout.actual
 !src/__tests__/domain-redirect.integ.cdkout
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/asset.*
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/**/asset.*
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/cdk.out
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/**/cdk.out
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/manifest.json
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/**/manifest.json
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/tree.json
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/**/tree.json
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/.cache
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout/**/.cache
-src/__tests__/backend/deny-list/integ/.tmp.deny-list.integ.cdkout.deploy
-src/__tests__/backend/deny-list/integ/.tmp.deny-list.integ.cdkout.actual
-!src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout
 !/src/spdx-license.ts
 !/.projenrc.ts
