+++ .npmignore
@@ -24,14 +24,12 @@ tsconfig.tsbuildinfo
 /lib/__tests__/
 /.eslintrc.json
 !.jsii
+/test
 !/website
 /update-github-ip-allowlist.js
 src/__tests__/.tmp.domain-redirect.integ.cdkout.deploy
 src/__tests__/.tmp.domain-redirect.integ.cdkout.actual
 src/__tests__/domain-redirect.integ.cdkout
-src/__tests__/backend/deny-list/integ/.tmp.deny-list.integ.cdkout.deploy
-src/__tests__/backend/deny-list/integ/.tmp.deny-list.integ.cdkout.actual
-src/__tests__/backend/deny-list/integ/deny-list.integ.cdkout
 /.gitattributes
 /.projenrc.ts
 /projenrc
