+++ .prettierignore
@@ -22,9 +22,9 @@ src/backend/ecs-task-monitor/monitor.ts
 src/backend/deny-list/prune-queue-handler.ts
 src/backend/deny-list/prune-handler.ts
 src/backend/catalog-builder/catalog-builder.ts
-src/__tests__/backend/deny-list/integ/trigger.prune-test.ts
-src/__tests__/backend/deny-list/integ/trigger.client-test.ts
-src/__tests__/backend/deny-list/integ/catalog-builder-mock.ts
+src/__tests__/backend/deny-list/mocks/trigger.prune-test.ts
+src/__tests__/backend/deny-list/mocks/trigger.client-test.ts
+src/__tests__/backend/deny-list/mocks/catalog-builder-mock.ts
 src/backend/transliterator/transliterator.ts
 src/backend/transliterator/transliterator.ecs-entrypoint.ts
 lib/backend/transliterator/transliterator.ecs-entrypoint.bundle/Dockerfile
