+++ package.json
@@ -79,10 +79,6 @@
     "eject": "npx projen eject",
     "eslint": "npx projen eslint",
     "integ": "npx projen integ",
-    "integ:deny-list:assert": "npx projen integ:deny-list:assert",
-    "integ:deny-list:deploy": "npx projen integ:deny-list:deploy",
-    "integ:deny-list:destroy": "npx projen integ:deny-list:destroy",
-    "integ:deny-list:snapshot": "npx projen integ:deny-list:snapshot",
     "integ:domain-redirect:assert": "npx projen integ:domain-redirect:assert",
     "integ:domain-redirect:deploy": "npx projen integ:domain-redirect:deploy",
     "integ:domain-redirect:destroy": "npx projen integ:domain-redirect:destroy",
