+++ src/__tests__/backend/deny-list/mocks/catalog-builder-mock.lambda.ts
