+++ src/__tests__/backend/deny-list/mocks/trigger.prune-test.lambda.ts
