+++ test/fixtures/deny-list/package-data/data/mypackage/v.1.0.0/file1.txt
