+++ test/fixtures/deny-list/package-data/data/mypackage/v1.2.0/foo.xoo
