+++ test/fixtures/deny-list/package-data/data/your/v1.2.4/world.txt
