+++ test/fixtures/transliterator/package/assembly.json
