+++ test/integ.deny-list.ts.snapshot/DenyListInteg.assets.json
@@ -92,28 +92,28 @@
         }
       }
     },
-    "a2b6fc284e07f2948c5e5664c1dfcff68ed9bbaef106822e51b3ffe723fce733": {
+    "48cfb47ac396c48b9fe4f0abbce5c930416e1465d1411aa3d1086b77f2b1e3de": {
       "source": {
-        "path": "asset.a2b6fc284e07f2948c5e5664c1dfcff68ed9bbaef106822e51b3ffe723fce733.bundle",
+        "path": "asset.48cfb47ac396c48b9fe4f0abbce5c930416e1465d1411aa3d1086b77f2b1e3de.bundle",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "a2b6fc284e07f2948c5e5664c1dfcff68ed9bbaef106822e51b3ffe723fce733.zip",
+          "objectKey": "48cfb47ac396c48b9fe4f0abbce5c930416e1465d1411aa3d1086b77f2b1e3de.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "0e3abd4bef72fe3e08d8e7354343c619146d3662fb87eacc9712b78f76338ab9": {
+    "6000e69f7aafe4d52f156ed4c9d619c21c66b4429ab5a024defaebfb6700044f": {
       "source": {
-        "path": "asset.0e3abd4bef72fe3e08d8e7354343c619146d3662fb87eacc9712b78f76338ab9.bundle",
+        "path": "asset.6000e69f7aafe4d52f156ed4c9d619c21c66b4429ab5a024defaebfb6700044f.bundle",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "0e3abd4bef72fe3e08d8e7354343c619146d3662fb87eacc9712b78f76338ab9.zip",
+          "objectKey": "6000e69f7aafe4d52f156ed4c9d619c21c66b4429ab5a024defaebfb6700044f.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -131,28 +131,28 @@
         }
       }
     },
-    "0371bf938ff2cb7d107f016044b2dd982a9a51c34445cf560faae19362254ebd": {
+    "48d993869418a3dd49aa6559ab0ebdd0dd35dc070685d051d22e14c67485a75b": {
       "source": {
-        "path": "asset.0371bf938ff2cb7d107f016044b2dd982a9a51c34445cf560faae19362254ebd.bundle",
+        "path": "asset.48d993869418a3dd49aa6559ab0ebdd0dd35dc070685d051d22e14c67485a75b.bundle",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "0371bf938ff2cb7d107f016044b2dd982a9a51c34445cf560faae19362254ebd.zip",
+          "objectKey": "48d993869418a3dd49aa6559ab0ebdd0dd35dc070685d051d22e14c67485a75b.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "1e2a72194021b2c99d55f08fdd8dc9e626a4982598931cfb75893a25d4d773fa": {
+    "3493864ccb154186e013addf6c6cc9d5f5811c0ea6b078d3d2d4507dc4c43371": {
       "source": {
-        "path": "TestDenyList.template.json",
+        "path": "DenyListInteg.template.json",
         "packaging": "file"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "1e2a72194021b2c99d55f08fdd8dc9e626a4982598931cfb75893a25d4d773fa.json",
+          "objectKey": "3493864ccb154186e013addf6c6cc9d5f5811c0ea6b078d3d2d4507dc4c43371.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
