+++ test/integ.deny-list.ts.snapshot/cdk.out
@@ -1,0 +1,1 @@
+{"version":"36.0.0"}
