+++ test/integ.deny-list.ts.snapshot/integ.json
@@ -1,0 +1,12 @@
+{
+  "version": "36.0.0",
+  "testCases": {
+    "deny-list-integ/DefaultTest": {
+      "stacks": [
+        "DenyListInteg"
+      ],
+      "assertionStack": "deny-list-integ/DefaultTest/DeployAssert",
+      "assertionStackName": "denylistintegDefaultTestDeployAssert070824D9"
+    }
+  }
+}
