+++ test/integ.transliterator.ecstask.ts.snapshot/TransliteratorEcsTaskInteg.assets.json
@@ -40,7 +40,7 @@
         }
       }
     },
-    "d629183b3b5a048fa164c7601cf0c01fe6a87cc1b7c2535a08a8413e2708ab95": {
+    "925ac73000275c52b8ba1927771a95d4e7bbd32bb4032463e9c96ab3710c7c47": {
       "source": {
         "path": "TransliteratorEcsTaskInteg.template.json",
         "packaging": "file"
@@ -48,22 +48,22 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "d629183b3b5a048fa164c7601cf0c01fe6a87cc1b7c2535a08a8413e2708ab95.json",
+          "objectKey": "925ac73000275c52b8ba1927771a95d4e7bbd32bb4032463e9c96ab3710c7c47.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     }
   },
   "dockerImages": {
-    "94f2518f19be80aed25d2146b42522cce8c3b71210c12d3f8b9eeef608433e8b": {
+    "8d872715ca58e94d53d354eb8a3b9d23c8717ecde82951789fbf0af223f31d2c": {
       "source": {
-        "directory": "asset.94f2518f19be80aed25d2146b42522cce8c3b71210c12d3f8b9eeef608433e8b.bundle",
+        "directory": "asset.8d872715ca58e94d53d354eb8a3b9d23c8717ecde82951789fbf0af223f31d2c.bundle",
         "platform": "linux/arm64"
       },
       "destinations": {
         "current_account-current_region": {
           "repositoryName": "cdk-hnb659fds-container-assets-${AWS::AccountId}-${AWS::Region}",
-          "imageTag": "94f2518f19be80aed25d2146b42522cce8c3b71210c12d3f8b9eeef608433e8b",
+          "imageTag": "8d872715ca58e94d53d354eb8a3b9d23c8717ecde82951789fbf0af223f31d2c",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-image-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
