+++ test/integ.transliterator.ecstask.ts.snapshot/TransliteratorEcsTaskInteg.template.json
@@ -1493,7 +1493,7 @@
       ],
       "Essential": true,
       "Image": {
-       "Fn::Sub": "${AWS::AccountId}.dkr.ecr.${AWS::Region}.${AWS::URLSuffix}/cdk-hnb659fds-container-assets-${AWS::AccountId}-${AWS::Region}:94f2518f19be80aed25d2146b42522cce8c3b71210c12d3f8b9eeef608433e8b"
+       "Fn::Sub": "${AWS::AccountId}.dkr.ecr.${AWS::Region}.${AWS::URLSuffix}/cdk-hnb659fds-container-assets-${AWS::AccountId}-${AWS::Region}:8d872715ca58e94d53d354eb8a3b9d23c8717ecde82951789fbf0af223f31d2c"
       },
       "LogConfiguration": {
        "LogDriver": "awslogs",
