+++ test/integ.transliterator.ecstask.ts.snapshot/transliteratorintegDefaultTestDeployAssert109DC8B0.assets.json
@@ -14,7 +14,7 @@
         }
       }
     },
-    "d1c573df1179c0a28edfc45dc8da18afec8fceb6ffcfd0bcaea00a1fde15e24a": {
+    "2de0f273b38ae0a962ca2852e54e66556eaa00947a786d03defe15d56942ef22": {
       "source": {
         "path": "transliteratorintegDefaultTestDeployAssert109DC8B0.template.json",
         "packaging": "file"
@@ -22,7 +22,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "d1c573df1179c0a28edfc45dc8da18afec8fceb6ffcfd0bcaea00a1fde15e24a.json",
+          "objectKey": "2de0f273b38ae0a962ca2852e54e66556eaa00947a786d03defe15d56942ef22.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
