+++ test/integ.transliterator.ecstask.ts.snapshot/transliteratorintegDefaultTestDeployAssert109DC8B0.template.json
@@ -41,7 +41,7 @@
     "outputPaths": [
      "executionArn"
     ],
-    "salt": "1725976481239"
+    "salt": "1726229680285"
    },
    "UpdateReplacePolicy": "Delete",
    "DeletionPolicy": "Delete"
@@ -158,7 +158,7 @@
      }
     },
     "flattenResponse": "false",
-    "salt": "1725976481241"
+    "salt": "1726229680287"
    },
    "UpdateReplacePolicy": "Delete",
    "DeletionPolicy": "Delete"
