+++ tsconfig.dev.json
@@ -28,6 +28,7 @@
   "include": [
     "src/**/*.ts",
     "src/__tests__/**/*.ts",
+    "test/**/*.ts",
     ".projenrc.ts",
     "projenrc/**/*.ts"
   ],
